# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: Order API
  description: API REST para gerenciamento de pedidos
  version: 1.0.0

servers:
  - url: http://localhost:3000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ItemInput:
      type: object
      properties:
        idItem:
          type: string
          example: "2434"
        quantidadeItem:
          type: integer
          example: 1
        valorItem:
          type: number
          example: 1000

    OrderInput:
      type: object
      required: [numeroPedido, valorTotal, dataCriacao, items]
      properties:
        numeroPedido:
          type: string
          example: "v10089015vdb-01"
        valorTotal:
          type: number
          example: 10000
        dataCriacao:
          type: string
          format: date-time
          example: "2023-07-19T12:24:11.529Z"
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemInput'

    OrderOutput:
      type: object
      properties:
        orderId:
          type: string
        value:
          type: number
        creationDate:
          type: string
          format: date-time
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              price:
                type: number

paths:
  /auth/token:
    post:
      summary: Gera um token JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        200:
          description: Token gerado com sucesso
        401:
          description: Credenciais inválidas

  /order:
    post:
      summary: Cria um novo pedido
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        201:
          description: Pedido criado com sucesso
        400:
          description: Dados inválidos
        409:
          description: Pedido já existe

  /order/list:
    get:
      summary: Lista todos os pedidos
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderOutput'

  /order/{orderId}:
    get:
      summary: Busca pedido por ID
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderOutput'
        404:
          description: Pedido não encontrado

    put:
      summary: Atualiza um pedido
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        200:
          description: Pedido atualizado
        404:
          description: Pedido não encontrado

    delete:
      summary: Deleta um pedido
      tags: [Orders]
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Pedido deletado
        404:
          description: Pedido não encontrado